+++ packages/zosjobs/__tests__/__unit__/CancelJobs.unit.test.ts
@@ -12,11 +12,12 @@
 import { ZosmfRestClient } from "@zowe/core-for-zowe-sdk";
 import { ImperativeError } from "@zowe/imperative";
 import { CancelJobs, IJob } from "../../src";
+import { CancelJobsData } from "../__resources__/api/CancelJobsData";
 
 jest.mock("@zowe/core-for-zowe-sdk");
 
-const returnEmpty = async () => {
-    return;
+const returnCancelJobsDataAsync = async () => {
+    return CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC;
 };
 
 const mockErrorText = "My fake error for unit tests has this text - Cancel Jobs unit tests";
@@ -42,76 +43,105 @@ describe("Cancel Jobs unit tests", () => {
         "phase-name": "OUTPUT",
         "job-correlator": "mycorrelator"
     };
+
     describe("Positive tests", () => {
         it("should allow users to call cancelJob with correct parameters", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(returnEmpty);
+            ZosmfRestClient.putExpectJSON = jest.fn(returnCancelJobsDataAsync);
             let caughtError;
+            let response;
             try {
-                await CancelJobs.cancelJob(fakeSession, "MYJOB1", "JOB00001");
+                response = await CancelJobs.cancelJob(fakeSession, "MYJOB1", "JOB00001");
             } catch (error) {
                 caughtError = error;
             }
             expect(caughtError).toBeUndefined();
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC);
         });
 
         it("should allow users to call cancelJobForJob with correct parameters", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(returnEmpty);
+            ZosmfRestClient.putExpectJSON = jest.fn(returnCancelJobsDataAsync);
             let caughtError;
+            let response;
             try {
-                await CancelJobs.cancelJobForJob(fakeSession, fakeJob);
+                response = await CancelJobs.cancelJobForJob(fakeSession, fakeJob);
             } catch (error) {
                 caughtError = error;
             }
             expect(caughtError).toBeUndefined();
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC);
         });
 
         it("should allow users to call cancelJobForJob with correct parameters (with version 1_0)", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
+            ZosmfRestClient.putExpectJSON = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
                 expect(body).toMatchSnapshot();
-                return {};
+                return CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC;
             });
-            await CancelJobs.cancelJobForJob(fakeSession, fakeJob, "1.0");
+            const response = await CancelJobs.cancelJobForJob(fakeSession, fakeJob, "1.0");
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC);
         });
 
         it("should allow users to call cancelJobForJob with correct parameters (with version 2_0)", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
+            ZosmfRestClient.putExpectJSON = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
                 expect(body).toMatchSnapshot();
-                return {};
+                return CancelJobsData.SAMPLE_JOB_FEEDBACK_GOOD;
             });
-            await CancelJobs.cancelJobForJob(fakeSession, fakeJob, "2.0");
+            const response = await CancelJobs.cancelJobForJob(fakeSession, fakeJob, "2.0");
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_GOOD);
+        });
+
+        it("should allow users to call cancelJobForJob with correct parameters (with version 2_0) and get a bad response", async () => {
+            ZosmfRestClient.putExpectJSON = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
+                expect(body).toMatchSnapshot();
+                return CancelJobsData.SAMPLE_JOB_FEEDBACK_BAD;
+            });
+            const response = await CancelJobs.cancelJobForJob(fakeSession, fakeJob, "2.0");
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_BAD);
         });
 
         it("should allow users to call cancelJobCommon with correct parameters", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(returnEmpty);
+            ZosmfRestClient.putExpectJSON = jest.fn(returnCancelJobsDataAsync);
             let caughtError;
+            let response;
             try {
-                await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001" });
+                response = await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001" });
             } catch (error) {
                 caughtError = error;
             }
             expect(caughtError).toBeUndefined();
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC);
         });
 
         it("should allow users to call cancelJobCommon with correct parameters (with version 1_0)", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
+            ZosmfRestClient.putExpectJSON = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
                 expect(body).toMatchSnapshot();
-                return {};
+                return CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC;
             });
-            await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001", version: "1.0" });
+            const response = await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001", version: "1.0" });
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_ASYNC);
         });
 
         it("should allow users to call cancelJobCommon with correct parameters (with version 2_0)", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
+            ZosmfRestClient.putExpectJSON = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
                 expect(body).toMatchSnapshot();
-                return {};
+                return CancelJobsData.SAMPLE_JOB_FEEDBACK_GOOD;
             });
-            await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001", version: "2.0" });
+            const response = await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001", version: "2.0" });
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_GOOD);
+        });
+
+        it("should allow users to call cancelJobCommon with correct parameters (with version 2_0) and get a failed response", async () => {
+            ZosmfRestClient.putExpectJSON = jest.fn(async (session: any, resource: string, headers: any[], body: any) => {
+                expect(body).toMatchSnapshot();
+                return CancelJobsData.SAMPLE_JOB_FEEDBACK_BAD;
+            });
+            const response = await CancelJobs.cancelJobCommon(fakeSession, { jobname: "MYJOB1", jobid: "JOB00001", version: "2.0" });
+            expect(response).toEqual(CancelJobsData.SAMPLE_JOB_FEEDBACK_BAD);
         });
     });
 
     describe("Error handling tests - async/await", () => {
         it("should be able to catch errors from cancelJob with async/await syntax", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJob(fakeSession, "MYJOB1", "JOB0000");
@@ -124,7 +154,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should be able to catch errors from cancelJobForJob with async/await syntax", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJobForJob(fakeSession, fakeJob);
@@ -137,7 +167,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should be able to catch errors from cancelJobCommon with async/await syntax", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJobCommon(fakeSession, {
@@ -156,7 +186,7 @@ describe("Cancel Jobs unit tests", () => {
     describe("Error handling tests - Promise catch() syntax", () => {
         /* eslint-disable jest/no-done-callback */
         it("should be able to catch errors from cancelJob with Promise.catch() syntax", (done: any) => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             CancelJobs.cancelJob(fakeSession, "MYJOB1", "JOB0000").then(() => {
                 expect(".catch() should have been called").toEqual("test failed");
             }).catch((err) => {
@@ -168,7 +198,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should be able to catch errors from cancelJobForJob with Promise.catch() syntax", (done: any) => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             CancelJobs.cancelJobForJob(fakeSession, fakeJob)
                 .then(() => {
                     expect(".catch() should have been called").toEqual("test failed");
@@ -181,7 +211,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should be able to catch errors from cancelJobCommon with Promise.catch() syntax", (done: any) => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             CancelJobs.cancelJobCommon(fakeSession, {
                 jobname: "MYJOB1",
                 jobid: "JOB0001"
@@ -200,7 +230,7 @@ describe("Cancel Jobs unit tests", () => {
     describe("Parameter validation", () => {
 
         it("should reject calls to cancelJob that omit jobname", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJob(fakeSession, undefined, "JOB0000");
@@ -213,7 +243,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should reject calls to cancelJob that omit jobid", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJob(fakeSession, "MYJOB1", undefined);
@@ -226,7 +256,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should reject calls to cancelJobForJob that omit jobname", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             const badJob: IJob = JSON.parse(JSON.stringify(fakeJob));
             delete badJob.jobname;
@@ -241,7 +271,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should reject calls to cancelJobForJob that omit jobid", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             const badJob: IJob = JSON.parse(JSON.stringify(fakeJob));
             delete badJob.jobid;
@@ -256,7 +286,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should reject calls to cancelJobCommon that omit jobname from the parms object", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJobCommon(fakeSession, {
@@ -272,7 +302,7 @@ describe("Cancel Jobs unit tests", () => {
         });
 
         it("should reject calls to cancelJobCommon that omit jobid from the parms object", async () => {
-            ZosmfRestClient.putExpectString = jest.fn(throwImperativeError);
+            ZosmfRestClient.putExpectJSON = jest.fn(throwImperativeError);
             let err: Error | ImperativeError;
             try {
                 await CancelJobs.cancelJobCommon(fakeSession, {